backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.0
bcrypt==3.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==3.4.8
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.3
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
mccabe==0.7.0
mongoengine==0.29.1
more-itertools==10.5.0
motor==3.6.0
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
postmarker==1.0
pyasn1==0.6.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
PyJWT==1.7.1
pymongo==4.9.2
pytest==8.3.3
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
python-ulid==1.1.0
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
tomli==2.0.2
types-cffi==1.16.0.20240331
types-cryptography==3.3.23.2
types-pyOpenSSL==23.0.0.0
types-setuptools==75.3.0.20241105
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
